made some various changes
