+++ lib/components/footer/bottomNavigationTabs.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:off_top_mobile/login/login.dart';
 import 'package:off_top_mobile/recordingSession.dart';
-import 'package:off_top_mobile/settings_page.dart';
+import 'package:off_top_mobile/settings/settingsPage.dart';
 
 class BottomNavigationTabs extends StatelessWidget {
   const BottomNavigationTabs(this.homePage);
